+++ app/src/main/java/com/example/soccertask/data/source/Team.kt
@@ -1,0 +1,9 @@
+package com.example.soccertask.data.source
+
+data class Team(
+    val teamName: String,
+    var points: Int = 0,
+    var gamesWon: Int = 0,
+    var gamesLost: Int = 0,
+    var drawGames: Int = 0
+)
